Declare -count as NSInteger in BROrm.h
Tabit PAD patch, previously applied in-place in Pods/ (TAB-49236): the
header declared -(int)count while PAD call sites consume NSInteger. The
implementation is intentionally left unchanged, mirroring the exact
state the app has shipped with for years.

+++ BROrm/BROrm.h
@@ -50,7 +50,7 @@
 - (NSArray*)findMany;
 - (NSArray*)findManyAsDictionaries;
 
-- (int)count;
+- (NSInteger)count;
 
 - (void)rawQuery:(NSString *)query withParameters:(NSArray*)parameters;
 - (void)select:(NSString*)column as:(NSString*)alias;
